requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
aiofiles==23.2.0
aiohttp==3.9.0
aiosqlite==0.19.0
Pillow==10.1.0
numpy==1.24.3
scikit-learn==1.3.2
opencv-python==4.8.1.78
async-lru==2.0.4
redis==5.0.1
celery==5.3.4
pydantic==2.5.0
tenacity==8.2.3
python-multipart==0.0.6